% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-fns.R
\name{get.ffns}
\alias{get.ffns}
\title{Functions for continuous variables, written in Fortran}
\usage{
get.ffns()
}
\value{
List of functions Functions for continuous
variables, in Fortran
}
\description{
Functions for continuous variables, written in Fortran
}

