% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt-fns.R
\name{get.rfns}
\alias{get.rfns}
\title{Functions for continuous variables, written in R}
\usage{
get.rfns()
}
\value{
List of R functions for continuous variables
}
\description{
Functions for continuous variables, written in R
}

